refactor: index.tsx 파일 생성 및 export 구문으로 일괄 변경
